requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.0
httpx==0.27.0
python-dotenv==1.0.1
openai==1.30.1
python-telegram-bot==21.2
gspread==6.1.0
google-auth==2.30.0
tenacity==8.3.0
structlog==24.2.0
